requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.3
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
attrs==23.2.0
    # via aiohttp
cffi==1.16.0
    # via cryptography
click==8.1.7
    # via redis-om
cryptography==42.0.5
    # via
    #   types-pyopenssl
    #   types-redis
discord-py==2.3.2
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
hiredis==2.3.2
    # via redis-om
idna==3.6
    # via yarl
loguru==0.7.2
    # via -r requirements.in
more-itertools==9.1.0
    # via redis-om
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
peewee==3.17.1
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic==2.0.3
    # via redis-om
pydantic-core==2.3.0
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
python-ulid==1.1.0
    # via redis-om
redis==4.6.0
    # via redis-om
redis-om==0.2.1
    # via -r requirements.in
types-pyopenssl==24.0.0.20240228
    # via types-redis
types-redis==4.6.0.20240218
    # via redis-om
typing-extensions==4.10.0
    # via
    #   pydantic
    #   pydantic-core
    #   redis-om
yarl==1.9.4
    # via aiohttp
